requirements.txt
APScheduler==3.6.3
art==5.5
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
greenlet==1.1.2
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pafy==0.5.5
Pillow==9.1.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytube==12.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
samila==0.6
six==1.16.0
SQLAlchemy==1.4.35
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
